requirements.txt
python-socketio
requests
sanic
websocket-client